refactor: combine singulars into BoardSkills and CardDeck and move files around
